// src/app/shared/model/post.ts
export interface Ipost {
  title: string;
  body: string;
}

export interface Iusers {
  fname: string;
  lname: string;
  cont: string;
  email: string;
}
export interface Istd {
  fname: string;
  lname: string;
  contact: number;
  email: string;
}
export interface itEmplyee {
    name: string;
    gender: string;
    age: number;
    department: string;
    salary: number;
    yearsOfExperience: number;
}